+++ lib/super_producer/manager.ex
@@ -1,5 +1,5 @@
-defmodule RabbitMQStream.SuperPublisher.Manager do
-  alias RabbitMQStream.SuperPublisher
+defmodule RabbitMQStream.SuperProducer.Manager do
+  alias RabbitMQStream.SuperProducer
 
   use GenServer
 
@@ -9,23 +9,23 @@ defmodule RabbitMQStream.SuperPublisher.Manager do
 
   @impl true
   def init(opts \\ []) do
-    state = struct(SuperPublisher, opts)
+    state = struct(SuperProducer, opts)
 
     {:ok, state, {:continue, :start}}
   end
 
   @impl true
-  def handle_continue(:start, %SuperPublisher{} = state) do
+  def handle_continue(:start, %SuperProducer{} = state) do
     for partition <- 0..(state.partitions - 1) do
       {:ok, _pid} =
         DynamicSupervisor.start_child(
           state.dynamic_supervisor,
           {
-            RabbitMQStream.Publisher,
-            Keyword.merge(state.publisher_opts,
+            RabbitMQStream.Producer,
+            Keyword.merge(state.producer_opts,
               name: {:via, Registry, {state.registry, partition}},
               stream_name: "#{state.super_stream}-#{partition}",
-              publisher_module: state.publisher_module
+              producer_module: state.producer_module
             )
           }
         )